Create buttons
This little script will run a while loop waiting for the user to push the button A or C (the leftmost and rightmost buttons) of the M5STACK head module. After the user has pushed the button C the script will come to a "break" command and stopping the execution.